Splitting up the structure of the PyDough pipeline tests (#280)
Splitting up the pipeline tests into multiple files to separate multiple distinctions: correlated tests, TPC-H queries, custom queries with TPC-H data, defog.ai queries, and custom queries with defog.ai data. Also, added all of the remaining unit tests for the defog.ai Broker queries (several of which need to be skipped until a followup is done)
