+++ tests/test_pipeline_defog.py
@@ -8,14 +8,27 @@ import pandas as pd
 import pytest
 from defog_outputs import (
     defog_sql_text_broker_adv1,
+    defog_sql_text_broker_adv2,
     defog_sql_text_broker_adv3,
+    defog_sql_text_broker_adv4,
+    defog_sql_text_broker_adv5,
     defog_sql_text_broker_adv6,
+    defog_sql_text_broker_adv7,
+    defog_sql_text_broker_adv8,
+    defog_sql_text_broker_adv9,
+    defog_sql_text_broker_adv10,
     defog_sql_text_broker_adv11,
     defog_sql_text_broker_adv12,
+    defog_sql_text_broker_adv13,
+    defog_sql_text_broker_adv14,
     defog_sql_text_broker_adv15,
+    defog_sql_text_broker_adv16,
+    defog_sql_text_broker_basic1,
+    defog_sql_text_broker_basic2,
     defog_sql_text_broker_basic3,
     defog_sql_text_broker_basic4,
     defog_sql_text_broker_basic5,
+    defog_sql_text_broker_basic6,
     defog_sql_text_broker_basic7,
     defog_sql_text_broker_basic8,
     defog_sql_text_broker_basic9,
@@ -23,14 +36,27 @@ from defog_outputs import (
 )
 from defog_test_functions import (
     impl_defog_broker_adv1,
+    impl_defog_broker_adv2,
     impl_defog_broker_adv3,
+    impl_defog_broker_adv4,
+    impl_defog_broker_adv5,
     impl_defog_broker_adv6,
+    impl_defog_broker_adv7,
+    impl_defog_broker_adv8,
+    impl_defog_broker_adv9,
+    impl_defog_broker_adv10,
     impl_defog_broker_adv11,
     impl_defog_broker_adv12,
+    impl_defog_broker_adv13,
+    impl_defog_broker_adv14,
     impl_defog_broker_adv15,
+    impl_defog_broker_adv16,
+    impl_defog_broker_basic1,
+    impl_defog_broker_basic2,
     impl_defog_broker_basic3,
     impl_defog_broker_basic4,
     impl_defog_broker_basic5,
+    impl_defog_broker_basic6,
     impl_defog_broker_basic7,
     impl_defog_broker_basic8,
     impl_defog_broker_basic9,
@@ -79,6 +105,14 @@ def test_graph_structure_defog(defog_graphs: graph_fetcher, graph_name: str) -> 
         ),
         pytest.param(
             (
+                impl_defog_broker_adv2,
+                "Broker",
+                defog_sql_text_broker_adv2,
+            ),
+            id="broker_adv2",
+        ),
+        pytest.param(
+            (
                 impl_defog_broker_adv3,
                 "Broker",
                 defog_sql_text_broker_adv3,
@@ -87,11 +121,68 @@ def test_graph_structure_defog(defog_graphs: graph_fetcher, graph_name: str) -> 
         ),
         pytest.param(
             (
+                impl_defog_broker_adv4,
+                "Broker",
+                defog_sql_text_broker_adv4,
+            ),
+            id="broker_adv4",
+        ),
+        pytest.param(
+            (
+                impl_defog_broker_adv5,
+                "Broker",
+                defog_sql_text_broker_adv5,
+            ),
+            id="broker_adv5",
+            marks=pytest.mark.skip(
+                "TODO (gh #158): add PREV function support to PyDough"
+            ),
+        ),
+        pytest.param(
+            (
                 impl_defog_broker_adv6,
                 "Broker",
                 defog_sql_text_broker_adv6,
             ),
             id="broker_adv6",
+        ),
+        pytest.param(
+            (
+                impl_defog_broker_adv7,
+                "Broker",
+                defog_sql_text_broker_adv7,
+            ),
+            id="broker_adv7",
+        ),
+        pytest.param(
+            (
+                impl_defog_broker_adv8,
+                "Broker",
+                defog_sql_text_broker_adv8,
+            ),
+            id="broker_adv8",
+            marks=pytest.mark.skip(
+                "TODO (gh #271): add 'week' support to PyDough DATETIME function"
+            ),
+        ),
+        pytest.param(
+            (
+                impl_defog_broker_adv9,
+                "Broker",
+                defog_sql_text_broker_adv9,
+            ),
+            id="broker_adv9",
+            marks=pytest.mark.skip(
+                "TODO (gh #271): add 'week' support to PyDough DATETIME function and DAYOFWEEK and WEEKOFYEAR functions"
+            ),
+        ),
+        pytest.param(
+            (
+                impl_defog_broker_adv10,
+                "Broker",
+                defog_sql_text_broker_adv10,
+            ),
+            id="broker_adv10",
         ),
         pytest.param(
             (
@@ -111,11 +202,51 @@ def test_graph_structure_defog(defog_graphs: graph_fetcher, graph_name: str) -> 
         ),
         pytest.param(
             (
+                impl_defog_broker_adv13,
+                "Broker",
+                defog_sql_text_broker_adv13,
+            ),
+            id="broker_adv13",
+        ),
+        pytest.param(
+            (
+                impl_defog_broker_adv14,
+                "Broker",
+                defog_sql_text_broker_adv14,
+            ),
+            id="broker_adv14",
+        ),
+        pytest.param(
+            (
                 impl_defog_broker_adv15,
                 "Broker",
                 defog_sql_text_broker_adv15,
             ),
             id="broker_adv15",
+        ),
+        pytest.param(
+            (
+                impl_defog_broker_adv16,
+                "Broker",
+                defog_sql_text_broker_adv16,
+            ),
+            id="broker_adv16",
+        ),
+        pytest.param(
+            (
+                impl_defog_broker_basic1,
+                "Broker",
+                defog_sql_text_broker_basic1,
+            ),
+            id="broker_basic1",
+        ),
+        pytest.param(
+            (
+                impl_defog_broker_basic2,
+                "Broker",
+                defog_sql_text_broker_basic2,
+            ),
+            id="broker_basic2",
         ),
         pytest.param(
             (
@@ -140,6 +271,14 @@ def test_graph_structure_defog(defog_graphs: graph_fetcher, graph_name: str) -> 
                 defog_sql_text_broker_basic5,
             ),
             id="broker_basic5",
+        ),
+        pytest.param(
+            (
+                impl_defog_broker_basic6,
+                "Broker",
+                defog_sql_text_broker_basic6,
+            ),
+            id="broker_basic6",
         ),
         pytest.param(
             (
@@ -179,7 +318,7 @@ def defog_test_data(
     request,
 ) -> tuple[Callable[[], UnqualifiedNode], Callable[[], str]]:
     """
-    Test data for test_defog_e2e. Returns a tuple of the following
+    Test data for `test_defog_e2e`. Returns a tuple of the following
     arguments:
     1. `unqualified_impl`: a PyDough implementation function.
     2. `graph_name`: the name of the graph from the defog database to use.
@@ -198,7 +337,7 @@ def test_defog_e2e(
     """
     Test executing the defog analytical questions on the sqlite database,
     comparing against the result of running the reference SQL query text on the
-    same database connector.
+    same database connector. Run on the defog.ai queries.
     """
     unqualified_impl, graph_name, query_impl = defog_test_data
     graph: GraphMetadata = defog_graphs(graph_name)
