+++ tests/test_sql_refsols/datetime_sampler_sqlite.sql
@@ -20,36 +20,33 @@ SELECT
   DATETIME('now') AS _expr18,
   DATETIME(order_date) AS _expr19,
   STRFTIME('%Y-%m-%d %H:%M:%S', DATETIME('now')) AS _expr20,
-  DATETIME('now', 'start of year', '8 minute', '-141 month') AS _expr21,
-  STRFTIME(
-    '%Y-%m-%d %H:00:00',
-    STRFTIME('%Y-%m-%d %H:%M:%S', DATETIME('now', 'start of month'))
-  ) AS _expr22,
+  DATETIME(DATE('now', 'start of year'), '8 minute', '-141 month') AS _expr21,
+  STRFTIME('%Y-%m-%d %H:00:00', STRFTIME('%Y-%m-%d %H:%M:%S', DATE('now', 'start of month'))) AS _expr22,
   STRFTIME(
     '%Y-%m-%d %H:00:00',
     STRFTIME('%Y-%m-%d %H:%M:%S', STRFTIME('%Y-%m-%d %H:00:00', DATETIME('now')))
   ) AS _expr23,
   DATETIME('now', '-96 hour', '15 year') AS _expr24,
-  DATETIME(
-    STRFTIME('%Y-%m-%d %H:%M:00', DATETIME('now', 'start of year', '-3 year')),
-    '65 month'
-  ) AS _expr25,
-  DATETIME(order_date, '-56 hour', 'start of year') AS _expr26,
+  DATETIME(STRFTIME('%Y-%m-%d %H:%M:00', DATE('now', 'start of year', '-3 year')), '65 month') AS _expr25,
+  DATE(DATETIME(order_date, '-56 hour'), 'start of year') AS _expr26,
   STRFTIME('%Y-%m-%d %H:%M:%S', STRFTIME('%Y-%m-%d %H:%M:00', DATETIME('now', '-63 day'))) AS _expr27,
-  DATETIME('now', 'start of month') AS _expr28,
+  DATE('now', 'start of month') AS _expr28,
   DATETIME(STRFTIME('%Y-%m-%d %H:%M:%S', DATETIME('now', '-312 hour')), '48 year') AS _expr29,
-  DATETIME('now', '75 day', 'start of day', '600 minute', '-294 day') AS _expr30,
-  DATETIME('now', 'start of month', '480 month', '-45 year') AS _expr31,
-  STRFTIME('%Y-%m-%d %H:%M:%S', DATETIME('now', '-270 minute', '-34 second', 'start of day')) AS _expr32,
-  DATETIME('now', 'start of month', '213 second') AS _expr33,
-  DATETIME('now', 'start of month', '13 minute', '28 year', '344 second') AS _expr34,
-  DATETIME('now', 'start of day') AS _expr35,
+  DATETIME(DATE(DATETIME('now', '75 day'), 'start of day'), '600 minute', '-294 day') AS _expr30,
+  DATE('now', 'start of month', '480 month', '-45 year') AS _expr31,
+  STRFTIME(
+    '%Y-%m-%d %H:%M:%S',
+    DATE(DATETIME('now', '-270 minute', '-34 second'), 'start of day')
+  ) AS _expr32,
+  DATETIME(DATE('now', 'start of month'), '213 second') AS _expr33,
+  DATETIME(DATE('now', 'start of month'), '13 minute', '28 year', '344 second') AS _expr34,
+  DATE('now', 'start of day') AS _expr35,
   DATETIME(STRFTIME('%Y-%m-%d %H:00:00', DATETIME('2025-01-01')), '49 minute', '91 year') AS _expr36,
-  DATETIME('now', 'start of year', 'start of day') AS _expr37,
-  DATETIME('now', 'start of day', 'start of year') AS _expr38,
-  DATETIME('2025-07-04 12:58:45', 'start of month', '22 minute') AS _expr39,
-  DATETIME('now', 'start of year') AS _expr40,
-  DATETIME(order_date, '82 second', '415 second', '-160 second', 'start of year') AS _expr41,
+  DATE('now', 'start of year', 'start of day') AS _expr37,
+  DATE('now', 'start of day', 'start of year') AS _expr38,
+  DATETIME(DATE('2025-07-04 12:58:45', 'start of month'), '22 minute') AS _expr39,
+  DATE('now', 'start of year') AS _expr40,
+  DATE(DATETIME(order_date, '82 second', '415 second', '-160 second'), 'start of year') AS _expr41,
   DATETIME('now', '192 month') AS _expr42,
   DATETIME(
     STRFTIME(
@@ -60,32 +57,30 @@ SELECT
   ) AS _expr43,
   DATETIME(STRFTIME('%Y-%m-%d %H:%M:%S', DATETIME('now')), '-50 hour') AS _expr44,
   STRFTIME('%Y-%m-%d %H:00:00', DATETIME('now', '297 day', '72 month', '-92 month')) AS _expr45,
-  DATETIME('now', '285 second', 'start of day') AS _expr46,
+  DATE(DATETIME('now', '285 second'), 'start of day') AS _expr46,
   DATETIME('1999-03-14', '62 day') AS _expr47,
-  DATETIME('now', 'start of month', '1 hour', 'start of month', '-21 day') AS _expr48,
+  DATE(DATETIME(DATE('now', 'start of month'), '1 hour'), 'start of month', '-21 day') AS _expr48,
   DATETIME('now', '212 minute', '368 year') AS _expr49,
   STRFTIME(
     '%Y-%m-%d %H:%M:00',
     STRFTIME(
       '%Y-%m-%d %H:%M:00',
-      DATETIME('2024-12-31 11:59:00', 'start of month', 'start of year')
+      DATE('2024-12-31 11:59:00', 'start of month', 'start of year')
     )
   ) AS _expr50,
-  DATETIME(STRFTIME('%Y-%m-%d %H:00:00', DATETIME('1999-03-14')), 'start of day') AS _expr51,
+  DATE(STRFTIME('%Y-%m-%d %H:00:00', DATETIME('1999-03-14')), 'start of day') AS _expr51,
   DATETIME(
-    STRFTIME('%Y-%m-%d %H:%M:00', DATETIME('now', '-60 hour', 'start of day')),
+    STRFTIME('%Y-%m-%d %H:%M:00', DATE(DATETIME('now', '-60 hour'), 'start of day')),
     '196 year'
   ) AS _expr52,
   DATETIME(STRFTIME('%Y-%m-%d %H:%M:00', DATETIME('now', '-40 hour', '-385 day')), '29 hour') AS _expr53,
   STRFTIME('%Y-%m-%d %H:%M:00', STRFTIME('%Y-%m-%d %H:00:00', DATETIME('now', '405 day'))) AS _expr54,
+  DATETIME(STRFTIME('%Y-%m-%d %H:%M:%S', DATE('now', 'start of year')), '98 year', '96 month') AS _expr55,
   DATETIME(
-    STRFTIME('%Y-%m-%d %H:%M:%S', DATETIME('now', 'start of year')),
-    '98 year',
-    '96 month'
-  ) AS _expr55,
-  DATETIME(
-    STRFTIME('%Y-%m-%d %H:%M:%S', STRFTIME('%Y-%m-%d %H:%M:00', DATETIME('now'))),
-    'start of day',
+    DATE(
+      STRFTIME('%Y-%m-%d %H:%M:%S', STRFTIME('%Y-%m-%d %H:%M:00', DATETIME('now'))),
+      'start of day'
+    ),
     '78 second'
   ) AS _expr56,
   DATETIME('now', '136 hour', '104 minute', '-104 month', '312 day') AS _expr57,
